% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notiondb-package.R
\docType{package}
\name{notiondb-package}
\alias{notiondb}
\alias{notiondb-package}
\title{notiondb: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/notiondb}
  \item Report bugs at \url{https://github.com/EthanSansom/notiondb/issues}
}

}
\author{
\strong{Maintainer}: Ethan Sansom \email{ethan.sansom29@gmail.com} (\href{https://orcid.org/0009-0000-1573-0186}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Marcelo Ponce
}

}
\keyword{internal}
