% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Print a description of the authenticated Notion user}
\usage{
whoami()
}
\value{
Nothing, called for side effects.
}
\description{
Print a description of the authenticated Notion user
}
\examples{
whoami()
}
